+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Added
+ - Input mask for default value for Textfield and Phone Number components.
+
 ## 2.27.1
 ### Fixed
  - Options to accept single-way binding.